+++ src/WebJobs.Script/Environment/EnvironmentSettingNames.cs
@@ -113,6 +113,7 @@ namespace Microsoft.Azure.WebJobs.Script
         public const string LinuxAzureAppServiceStorage = "WEBSITES_ENABLE_APP_SERVICE_STORAGE";
         public const string CoreToolsEnvironment = "FUNCTIONS_CORETOOLS_ENVIRONMENT";
         public const string RunningInContainer = "DOTNET_RUNNING_IN_CONTAINER";
+        public const string EnableCorsConfiguration = "FUNCTIONS_ENABLE_CORS_CONFIGURATION";
 
         public const string ExtensionBundleSourceUri = "FUNCTIONS_EXTENSIONBUNDLE_SOURCE_URI";